Code style unification
