+++ lib/components/Label.js
@@ -20,6 +20,7 @@ const Label = ({
     ...otherHelpIconProps
   } = helpIconProps || {};
   const HelpIcon = icon || Info;
+
   return (
     <label
       className={classnames("neeto-ui-label flex items-center", className)}
