+++ lib/components/Radio/index.js
@@ -2,8 +2,9 @@ import React from "react";
 import classnames from "classnames";
 import PropTypes from "prop-types";
 import { useId } from "@reach/auto-id";
-import { hyphenize } from "../common";
-import Label from "./Label";
+import { hyphenize } from "../../common";
+import Label from "../Label";
+import Item from "./Item";
 
 const Radio = ({
   label = "",
@@ -17,6 +18,7 @@ const Radio = ({
 }) => {
   const id = useId(props.id);
   const errorId = `error_${id}`;
+
   return (
     <div className={classnames(["neeto-ui-radio__wrapper", className])}>
       {label && <Label className="neeto-ui-radio__label">{label}</Label>}
@@ -72,27 +74,6 @@ Radio.propTypes = {
    * To specify the error message to be displayed when the radio input is invalid.
    */
   error: PropTypes.string,
-};
-
-const Item = ({ name = "", label = "", className = "", ...otherProps }) => {
-  const id = useId(otherProps.id);
-
-  return (
-    <div className={classnames(["neeto-ui-radio__item", className])}>
-      <input
-        id={id}
-        type="radio"
-        name={name}
-        className="neeto-ui-radio"
-        {...otherProps}
-      />
-      {label && (
-        <Label data-cy={`${hyphenize(label)}-radio-label`} htmlFor={id}>
-          {label}
-        </Label>
-      )}
-    </div>
-  );
 };
 
 Radio.Item = Item;
