+++ lib/components/Select.js
@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React from "react";
 import classnames from "classnames";
 import PropTypes from "prop-types";
 import SelectInput, { components } from "react-select";
@@ -27,81 +27,71 @@ const CustomInput = (props) => {
   );
 };
 
-class Select extends Component {
-  render() {
-    const {
-      size = sizes.large,
-      label = "",
-      required = false,
-      error = null,
-      helpText = "",
-      className = "",
-      innerRef,
-      isCreateable,
-      strategy = strategies.default,
-      ...otherProps
-    } = this.props;
+const Select = ({
+  size = sizes.large,
+  label = "",
+  required = false,
+  error = null,
+  helpText = "",
+  className = "",
+  innerRef,
+  isCreateable,
+  strategy = strategies.default,
+  ...otherProps
+}) => {
+  let Parent = SelectInput;
 
-    let Parent = SelectInput;
-
-    if (isCreateable) {
-      Parent = AsyncCreatable;
-    }
-
-    const portalProps =
-      strategy === strategies.fixed
-        ? {
-          menuPortalTarget: document.body,
-          styles: { menuPortal: (base) => ({ ...base, zIndex: 999999 }) },
-          menuPosition: "fixed",
-        }
-        : {};
-
-    return (
-      <div
-        className={classnames(["neeto-ui-input__wrapper", className])}
-        data-cy={`${hyphenize(label)}-select-container-wrapper`}
-      >
-        {label && (
-          <Label
-            required={required}
-            data-cy={`${hyphenize(label)}-input-label`}
-          >
-            {label}
-          </Label>
-        )}
-        <Parent
-          ref={innerRef}
-          className={classnames(["neeto-ui-react-select__container"], {
-            "neeto-ui-react-select__container--error": !!error,
-            "neeto-ui-react-select__container--small": size === sizes.small,
-          })}
-          classNamePrefix="neeto-ui-react-select"
-          components={{ Input: CustomInput }}
-          data-cy={`${hyphenize(label)}-select-container`}
-          {...portalProps}
-          {...otherProps}
-        />
-        {!!error && (
-          <p
-            className="neeto-ui-input__error"
-            data-cy={`${hyphenize(label)}-select-error`}
-          >
-            {error}
-          </p>
-        )}
-        {helpText && (
-          <p
-            className="neeto-ui-input__help-text"
-            data-cy={`${hyphenize(label)}-select-help-text`}
-          >
-            {helpText}
-          </p>
-        )}
-      </div>
-    );
+  if (isCreateable) {
+    Parent = AsyncCreatable;
   }
-}
+
+  const portalProps = strategy === strategies.fixed && {
+    menuPortalTarget: document.body,
+    styles: { menuPortal: (base) => ({ ...base, zIndex: 999999 }) },
+    menuPosition: "fixed",
+  };
+
+  return (
+    <div
+      className={classnames(["neeto-ui-input__wrapper", className])}
+      data-cy={`${hyphenize(label)}-select-container-wrapper`}
+    >
+      {label && (
+        <Label required={required} data-cy={`${hyphenize(label)}-input-label`}>
+          {label}
+        </Label>
+      )}
+      <Parent
+        ref={innerRef}
+        className={classnames(["neeto-ui-react-select__container"], {
+          "neeto-ui-react-select__container--error": !!error,
+          "neeto-ui-react-select__container--small": size === sizes.small,
+        })}
+        classNamePrefix="neeto-ui-react-select"
+        components={{ Input: CustomInput }}
+        data-cy={`${hyphenize(label)}-select-container`}
+        {...portalProps}
+        {...otherProps}
+      />
+      {!!error && (
+        <p
+          className="neeto-ui-input__error"
+          data-cy={`${hyphenize(label)}-select-error`}
+        >
+          {error}
+        </p>
+      )}
+      {helpText && (
+        <p
+          className="neeto-ui-input__help-text"
+          data-cy={`${hyphenize(label)}-select-help-text`}
+        >
+          {helpText}
+        </p>
+      )}
+    </div>
+  );
+};
 
 Select.propTypes = {
   size: PropTypes.oneOf(Object.values(sizes)),
