+++ lib/components/Switch.js
@@ -17,6 +17,7 @@ const Switch = ({
   const { checked, disabled } = otherProps;
   const id = useId(otherProps.id);
   const errorId = `error_${id}`;
+
   return (
     <div className={classnames(["neeto-ui-switch__wrapper", className])}>
       <div className="neeto-ui-switch__container">
