+++ lib/components/Tab/Item.js
@@ -5,50 +5,6 @@ import classnames from "classnames";
 
 const noop = () => {};
 
-const TAB_SIZES = { large: "large", default: "default" };
-
-const Tab = ({ children, size, noUnderline, className = "", ...otherProps }) => {
-  return (
-    <div
-      className={classnames(
-        {
-          "neeto-ui-tab__wrapper flex": true,
-        },
-        {
-          "neeto-ui-tab__wrapper--size-large": size === TAB_SIZES.large,
-        },
-        {
-          "neeto-ui-tab__wrapper--underline-none": noUnderline,
-        },
-        [className]
-      )}
-      data-cy="tab-container"
-      {...otherProps}
-    >
-      {children}
-    </div>
-  );
-};
-
-Tab.propTypes = {
-  /**
-   * To hide the underline
-   */
-  noUnderline: PropTypes.bool,
-  /**
-   * Set the size of the tabs.
-   */
-  size: PropTypes.oneOf(Object.values(TAB_SIZES)),
-  /**
-   * To add content inside Tab
-   */
-  children: PropTypes.node,
-  /**
-   * Extra classes can be provided to the Tab
-   */
-  className: PropTypes.string,
-};
-
 const Item = ({
   active,
   className = "",
@@ -67,6 +23,7 @@ const Item = ({
   if (activeClassName) {
     Parent = NavLink;
   }
+
   return (
     <Parent
       className={classnames(
@@ -112,6 +69,4 @@ Item.propTypes = {
   activeClassName: PropTypes.string,
 };
 
-Tab.Item = Item;
-
-export default Tab;
+export default Item;
